+++ src/config/googlePlaces/index.js
@@ -1,5 +1,4 @@
 const googlePlaces = {
-  maxImageHeight: 200,
   apiKey: 'AIzaSyAu5M0b3zBi4gT50BjpspZnEtCNeyCJwvY',
 };